Initial DTO and service implementation - FE-41741
Copied resources over from core because otherwise swagger would not notice them and the API docs won't build properly.

It is now possible to use both java -jar and tomcat:run-war to run the application. For the war, the documentation is at the root of the context, e.g., localhost:8080/coevolution-service/ .

(Subtask: [GeoKnow] Finalize D4.3.1)